[refactor] #98 : 바텀시트 폴더 구분
