+++ src/features/mapView/ui/bottomsheets/SnapBottomSheet.tsx
@@ -1,5 +1,5 @@
 import { SnapBottomSheet } from "@/shared/ui";
-import { BottomSheetContent, FixedButtons, GroupAverageTime } from "../durationBottomsheet";
+import { BottomSheetContent, FixedButtons, GroupAverageTime } from "./durationBottomsheet";
 
 export const SnapMapBottomSheet = () => {
   return (
