+++ src/features/mapView/ui/bottomsheets/detail/TransferDetail.tsx
