+++ src/features/mapView/ui/bottomsheets/detail/index.ts
