+++ src/features/mapView/ui/bottomsheets/detailContents/Path.tsx
