+++ src/features/mapView/ui/bottomsheets/detailContents/WalkPath.tsx
