+++ src/features/mapView/ui/bottomsheets/detailContents/index.ts
