+++ src/features/mapView/ui/bottomsheets/durationBottomsheet/BottomSheetContent.tsx
@@ -5,7 +5,7 @@ import { useNavigate, useSearchParams } from "react-router-dom";
 import AddUser from "@/assets/icon/addUser.svg";
 import Share from "@/assets/icon/share.svg";
 import { UserCard } from "../detail";
-import { LoginModal } from "../common";
+import { LoginModal } from "../..";
 
 export const BottomSheetContent = () => {
   const eventData = useEventStore(state => state.eventData);
